+++ crates/conversion/vec2svg/Cargo.toml
