+++ crates/conversion/vec2svg/README.md
