+++ crates/conversion/vec2svg/src/backend/glyph.rs
