+++ crates/conversion/vec2svg/src/backend/text.rs
