+++ crates/conversion/vec2svg/src/frontend/context.rs
