+++ crates/conversion/vec2svg/src/frontend/dynamic_layout.rs
