+++ crates/conversion/vec2svg/src/frontend/flat.rs
