+++ crates/conversion/vec2svg/src/frontend/incremental.rs
