+++ crates/conversion/vec2svg/src/frontend/mod.rs
