+++ crates/conversion/vec2svg/src/frontend/typst.svg.js
