+++ crates/conversion/vec2svg/src/lib.rs
