+++ crates/conversion/vec2svg/src/transform/minify.rs
