+++ crates/conversion/vec2svg/src/utils.rs
