--- exporter/ast/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "typst-ts-ast-exporter"
-description = "Dump Typst ast to specific format."
-authors.workspace = true
-version.workspace = true
-license.workspace = true
-edition.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[dependencies]
-
-typst.workspace = true
-
-ansi_term.workspace = true
