--- exporter/ast/README.md
@@ -1,3 +1,0 @@
-# typst-ts-ast-exporter
-
-See [Typst.ts](https://github.com/Myriad-Dreamin/typst.ts)
