+++ packages/compiler/Cargo.toml
@@ -46,7 +46,6 @@ typst-ts-core = { workspace = true, features = ["pdf", "svg"] }
 typst-ts-compiler = { version = "0.5.0-rc6", default-features = false, features = [
     "web",
 ] }
-typst-ts-ast-exporter.workspace = true
 reflexo-vec2canvas.workspace = true
 typst-ts-svg-exporter.workspace = true
